Fix generics resolution
