+++ tests/test__describe.py
@@ -663,6 +663,7 @@ def test_describe__tagged_union():
         custom_encoder=None,
     )
 
+
 @pytest.mark.skip(reason="https://github.com/python/cpython/pull/111515#issuecomment-2018132920")
 def test_describe__typed_dict():
     class Entity(TypedDict, Generic[T]):
